Refactors Android plugin alongside the embedding V2 migration
